Move SunAndMoonAnimatedSvg into its own component and tidy up
