+++ src/index.tsx
@@ -1,3 +1,3 @@
-export { DarkModeSwitch } from './react-toggle-dark-mode';
-export { defaultProperties } from './react-toggle-dark-mode';
+export { DarkModeSwitch } from './DarkModeSwitch';
+export { defaultProperties } from './DarkModeSwitch';
 export { ThemeMode, ColorOptions } from './types';
